worked on user endpoint
