+++ README.md
@@ -1,5 +1,9 @@
 # The first fighting themed recipe website!
 
-### For bodyguards by bodyguards!
+### For bodyguards by bodyguards<sup>*</sup>!
 
-Visit at [recipow.com](https://recipow.com)
+Visit at [recipow.com](https://recipow.com)<sup>**</sup>
+
+###### <sup>*</sup>Website is not made by bodyguards
+
+###### <sup>**</sup>Website is in early development and any data you upload should be considered unsafe/ unstable.
