+++ package.json
@@ -17,30 +17,29 @@
 		"@sveltejs/adapter-auto": "next",
 		"@sveltejs/adapter-netlify": "^1.0.0-next.64",
 		"@sveltejs/kit": "next",
-		"@types/cookie": "^0.4.1",
+		"@types/cookie": "^0.5.1",
 		"@types/debug": "^4.1.7",
-		"@types/ioredis": "^4.28.8",
+		"@types/ioredis": "^4.28.10",
 		"@types/string-hash": "^1.1.1",
 		"@types/uuid": "^8.3.4",
-		"@typescript-eslint/eslint-plugin": "^5.14.0",
-		"@typescript-eslint/parser": "^5.14.0",
-		"debug": "^4.3.3",
-		"eslint": "^7.32.0",
-		"eslint-config-prettier": "^8.3.0",
-		"eslint-plugin-svelte3": "^3.4.1",
-		"mongodb": "^4.5.0",
-		"prettier": "^2.5.1",
-		"prettier-plugin-svelte": "^2.5.0",
+		"@typescript-eslint/eslint-plugin": "^5.28.0",
+		"@typescript-eslint/parser": "^5.28.0",
+		"debug": "^4.3.4",
+		"eslint": "^8.17.0",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-plugin-svelte3": "^4.0.0",
+		"mongodb": "^4.7.0",
+		"prettier": "^2.7.0",
+		"prettier-plugin-svelte": "^2.7.0",
 		"redis-json": "^6.0.3",
-		"svelte": "^3.46.0",
-		"svelte-check": "^2.2.6",
-		"svelte-preprocess": "^4.10.1"
+		"svelte": "^3.48.0",
+		"svelte-check": "^2.7.2",
+		"svelte-preprocess": "^4.10.7"
 	},
 	"dependencies": {
-		"@fontsource/fira-mono": "^4.5.0",
-		"cookie": "^0.4.1",
-		"ioredis": "^4.28.5",
-		"redis": "^4.0.4",
+		"cookie": "^0.5.0",
+		"ioredis": "^5.0.6",
+		"redis": "^4.1.0",
 		"string-hash": "^1.1.3",
 		"uuid": "^8.3.2"
 	}
